Milestone_3/requirements.txt
# CardioDetect Milestone 3 - Requirements
# Django Backend with ML Integration

# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL adapter
django-redis>=5.3.0     # Redis cache backend

# Image/File handling
Pillow>=10.0.0

# ML Dependencies (from Milestone 2)
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# OCR Dependencies
pytesseract>=0.3.10
opencv-python>=4.8.0
PyMuPDF>=1.23.0

# Utilities
python-dotenv>=1.0.0

# Excel Export
openpyxl>=3.1.0

# PDF Generation
reportlab>=4.0.0

